GWSC.jl

GWSC.fm

GWSC.LISA

GWSC.PTA

GWSC.RLISA

GWSC.backup

GWSC.getSnC

GWSC.plotCharacteristicStrain

GWSC.plotSpectralDensity

GWSC.LISA — Type

LISA type

TObs::Float: observation time of LISA (yr);

LArm: Arm length of for every arm of LISA (meter);

NC: number of data channels for LISA;

approQ=false: whether to approximate the transfer function or not;

fStar::Float: transfer frequency (Hz)

R::Function: transfer function

Pn::Function: Power Spectral Density without confusion (WC) noise

Sn::Function: Sensitivity curve of LISA without the galactic binary confusion noise

Pn_WC::Function: Power Spectral Density with confusion (WC) noise

Sn_WC::Function: Sensitivity curve of LISA with the galactic binary confusion noise

SnC::Function: galactic binary confusion noise

Ωn::Function: effective energy density of noise

GWSC.PTA — Type

PTA <: Detector

Detector type for pulsar timing array (PTA).

Fields

Δt::Float=14DAY: (s); 1/Δt is the cadence of observations

σRMS::Float=1e2: root-mean-square timing noise (ns)

NP::Int=20: number of pulsars

TObs::Float=5.: observation time of PTA (yr)

fMin::Float: minimum frequency (Hz)

fMax::Float: maximum frequency (Hz)

fRef::Float=1e-8: reference frequency (Hz)

fPlotRange::Tuple{Float,Float}=(1e-9, 1e-7): frequency range to show (Hz, Hz)

ΩPlotRange::Tuple{Float,Float}=(1e-11, 1e-7): energy density range to show

hPlotRange::Tuple{Float,Float}=(1e-17, 1e-12): strain range to show

R::Function: transfer function

Pn::Function: detector power spectral density (Hz^-1)

Sn::Function: strain noise power spectral density (Hz^-1)

Seff::Function: effective strain noise power spectral density (Hz^-1)

Ωn::Function: fractional energy density of noise

Ωeff::Function: effective fractional energy density of noise

ρThSGWB::Float=1.0: threshold SNR for deteting a gravitational wave background

Examples

julia> pta = PTA(Δt=14*DAY, σRMS=1e2, NP=36, TObs=5.);

GWSC.backup — Method

backup(array1, array2, fileName::String)

Write the contents of two arrays to a file.

array1: the first column of the file;

array2: the second column of the file;

fileName: name of the file to be written to.

GWSC.plotCharacteristicStrain — Method

plotCharacteristicStrain(lisa::LISA; figurefile=undef, plotSnWC=true, plotSn=false, plotPn=false)

Plot the sensitivity curve of characteristic strain. If figure_file is provided, the figure will be saved.

lisa::LISA: lisa is a LISA type;

figure_file=undef: file name of the figue to be saved to;

plotSn_WC=true: plot √f*Sn_WC if it is true;

plotSn=false: plot √f*Sn if it is true;

plotPn=false: plot √f*Pn if it is true;

plotPn_WC=false: plot √f*Pn_WC if it is true;

GWSC.plotSpectralDensity — Method

plotSpectralDensity(lisa::LISA; figurefile=undef, plotSnWC=true, plotSn=false, plotPn=false, plotPn_WC=false)

Plot the sensitivity curve of characteristic strain. If figure_file is provided, the figure will be saved.

lisa::LISA: lisa is a LISA type;

figure_file=undef: file name of the figue to be saved to;

plotSn_WC=true: plot √Sn_WC if it is true;

plotSn=false: plot √Sn if it is true;

plotPn=false: plot √Pn if it is true;

plotPn_WC=false: plot √Pn_WC if it is true.

GWSC.fm — Constant

LISA modulation frequency

GWSC.RLISA — Function

RLISA(fStar::Float=19.09e-3, NC::Int=2; approQ=false)

Calculate the transfer function (sky and polarization averaged).

approQ=false: whether to approximate the transfer function or not;

fStar: transfer frequency;

NC: number of data channels for LISA.

GWSC.getSnC — Function

SnC(f::Float, TObs::Float=4.0, NC::Int=2)

An estimation of the galactic binary confusion noise for Tobs = {0.5 yr, 1 yr, 2 yr, 4yr}. Enter Tobs as a year or fraction of a year.

f: frequency (Hz);

TObs: observation time of LISA (yr);

NC: number of data channels for LISA.